+++ packages/documentation/src/utils/codeOrSourceMdx.tsx
@@ -3,6 +3,7 @@ import { Source } from '@storybook/blocks';
 import { Code } from '@storybook/components';
 import { SourceDarkMode } from '@/../.storybook/preview';
 
+/* eslint-disable-next-line @typescript-eslint/ban-ts-comment */
 // @ts-ignore
 export const CodeOrSourceMdx = ({ className, children, ...rest }) => {
   // markdown-to-jsx does not add className to inline code