+++ packages/documentation/src/utils/map-classes.ts
@@ -1,7 +1,6 @@
 export function mapClasses(classObject: Record<string, boolean>): string {
   return Object.entries(classObject)
-    .filter(([_newClass, shouldAddClass]) => shouldAddClass)
+    .filter(([_newClass, shouldAddClass]) => shouldAddClass && _newClass !== 'null')
     .map(([newClass]) => newClass)
-    .filter(c => c !== 'null')
     .join(' ');
 }