+++ packages/internet-header/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @swisspost/internet-header
 
+## 1.13.11
+
+### Patch Changes
+
+- Fixed focus trap on overlay of the breadcrumb. Added a focus trap on search overlay (by [@imagoiq](https://github.com/imagoiq) with [#2922](https://github.com/swisspost/design-system/pull/2922))
+- Updated dependencies:
+  - @swisspost/design-system-styles@7.2.0
+
 ## 1.13.10
 
 ### Patch Changes